Added GLib event loop support to wayland backend
